refactor methods to functions.js
